refactor: replace analyzer tabs with master-detail Results view
