--- ViewModels/Pages/PalletAnalyzerViewModel.cs
@@ -1,455 +1,0 @@
-using Stack_Solver.Helpers.Rendering;
-using Stack_Solver.Infrastructure;
-using Stack_Solver.Models;
-using Stack_Solver.Models.Assignment;
-using Stack_Solver.Models.Layering;
-using Stack_Solver.Models.Supports;
-using Stack_Solver.Services;
-using Stack_Solver.Services.Stacking;
-using System.Collections.ObjectModel;
-using System.Text;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Media3D;
-using Wpf.Ui;
-using Wpf.Ui.Controls;
-
-namespace Stack_Solver.ViewModels.Pages
-{
-    public partial class PalletAnalyzerViewModel : ObservableObject
-    {
-        private readonly IEventAggregator _events;
-        private readonly PalletSceneBuilder _sceneBuilder = new();
-        private CancellationTokenSource? _buildCts;
-        private CancellationTokenSource? _sceneCts;
-        private ViewportController? _viewportController;
-
-        private List<Layer> _availableLayers = [];
-        private int _palletLength, _palletWidth, _palletHeight;
-        private int _maxStackHeight = 180, _maxStackWeight = 950;
-        private double _maxSkuOverhang;
-        private List<SKU> _selectedSkus = [];
-        private GenerationOptions _generationOptions = new();
-
-        [ObservableProperty]
-        private bool _isBuilding;
-
-        [ObservableProperty]
-        private string _palletGenStats = "Click on 'Generate' to start solution generation.";
-
-        [ObservableProperty]
-        private ObservableCollection<TemplateAssignmentDisplay> _assignments = [];
-
-        [ObservableProperty]
-        private TemplateAssignmentDisplay? _selectedAssignment;
-
-        [ObservableProperty]
-        private bool _hasResults;
-
-        [ObservableProperty]
-        private bool _hasLayers;
-
-        [ObservableProperty]
-        private ObservableCollection<LayerTypeDisplay> _selectedLayerTypes = [];
-
-        [ObservableProperty]
-        private ObservableCollection<SolutionDisplay> _solutions = [];
-
-        [ObservableProperty]
-        private SolutionDisplay? _selectedSolution;
-
-        [ObservableProperty]
-        private string _assignmentDetailsText = string.Empty;
-
-        [ObservableProperty]
-        private LayerTypeDisplay? _selectedLayerType;
-
-        private readonly List<Model3DGroup> _layerHighlights = [];
-
-        public Model3DGroup Scene { get; } = new();
-        public IReadOnlyList<LabelInfo> PalletDimLabels => _sceneBuilder.DimLabels;
-        public ViewportController? ViewportController => _viewportController;
-        public ICommand ZoomCommand { get; }
-        public ICommand BeginPanCommand { get; }
-        public ICommand PanCommand { get; }
-
-        private readonly ISnackbarService _snackbarService;
-
-        public PalletAnalyzerViewModel(IEventAggregator events, ISnackbarService snackbarService)
-        {
-            _events = events;
-            _snackbarService = snackbarService;
-            _events.Subscribe<LayersGeneratedMessage>(OnLayersGenerated);
-            _events.Subscribe<SettingsChangedMessage>(OnSettingsChanged);
-            ZoomCommand = new RelayCommand<double>(delta => _viewportController?.Zoom(delta));
-            BeginPanCommand = new RelayCommand<Point>(p => _viewportController?.BeginPan(p));
-            PanCommand = new RelayCommand<Point>(p => _viewportController?.Pan(p));
-        }
-
-        public void AttachCamera(PerspectiveCamera camera)
-        {
-            if (camera == null) return;
-            if (_viewportController == null)
-            {
-                _viewportController = new ViewportController(camera, CurrentPalletCenter);
-                OnPropertyChanged(nameof(ViewportController));
-            }
-            else
-            {
-                _viewportController.Target = CurrentPalletCenter;
-            }
-        }
-
-        private Point3D CurrentPalletCenter => new(_palletLength / 2.0, 0, _palletWidth / 2.0);
-
-        private void OnLayersGenerated(LayersGeneratedMessage msg)
-        {
-            _availableLayers = msg.Layers;
-            HasLayers = _availableLayers.Count > 0;
-            HasResults = false;
-            Assignments.Clear();
-            Solutions.Clear();
-
-            _buildCts?.Cancel();
-            _buildCts?.Dispose();
-            _buildCts = new CancellationTokenSource();
-            _ = BuildPalletsAsync(_buildCts.Token);
-        }
-
-        private void OnSettingsChanged(SettingsChangedMessage msg)
-        {
-            _palletLength = msg.PalletLength;
-            _palletWidth = msg.PalletWidth;
-            _palletHeight = (int)Math.Round(msg.PalletHeight);
-            _maxStackHeight = msg.MaxStackHeight;
-            _maxStackWeight = msg.MaxStackWeight;
-            _maxSkuOverhang = msg.MaxSkuOverhang;
-            _selectedSkus = [.. msg.Skus.Where(s => s.IsSelected && s.Quantity > 0)];
-            _generationOptions = new GenerationOptions(msg.SolverTimeLimit, msg.MaxCpsatCandidates, msg.BlfAttempts);
-            _viewportController?.Target = CurrentPalletCenter;
-        }
-
-        private async Task BuildPalletsAsync(CancellationToken ct)
-        {
-            if (_availableLayers.Count == 0 || _selectedSkus.Count == 0)
-            {
-                PalletGenStats = _availableLayers.Count == 0
-                    ? "No layers available."
-                    : "No SKUs selected with quantity > 0.";
-                return;
-            }
-
-            IsBuilding = true;
-            try
-            {
-                var pallet = new Pallet("Pallet", _palletLength, _palletWidth, _palletHeight)
-                {
-                    MaxStackHeight = _maxStackHeight,
-                    MaxStackWeight = _maxStackWeight,
-                    MaxSkuOverhang = _maxSkuOverhang
-                };
-                var demand = _selectedSkus.ToDictionary(s => s.SkuId, s => s.Quantity, StringComparer.Ordinal);
-                var options = _generationOptions;
-                var skus = _selectedSkus.ToList();
-                var layersSnapshot = _availableLayers.ToList();
-
-                var greedyResult = await Task.Run(() =>
-                {
-                    ct.ThrowIfCancellationRequested();
-                    var filtered = LayerMetricsCalculator.FilterLayers(layersSnapshot, options);
-                    var greedy = GreedyAssignmentService.Assign(filtered, demand, pallet);
-                    if (!greedy.HasLeftovers) return greedy;
-
-                    var leftoverSkus = BuildLeftoverSkus(skus, greedy.Leftovers);
-                    if (leftoverSkus.Count == 0) return greedy;
-                    var tailLayers = LayerGenerator.Generate(leftoverSkus, pallet, options, ct: ct);
-                    if (tailLayers.Count == 0) return greedy;
-                    var tailOptions = new GenerationOptions(options.MaxSolverTime, options.MaxCPSATCandidates, options.BLFAttempts)
-                    {
-                        MaxLayerStability = options.MaxLayerStability,
-                        PerSkuTopLayerFraction = 1.0
-                    };
-                    var tailFiltered = LayerMetricsCalculator.FilterLayers(tailLayers, tailOptions);
-                    if (tailFiltered.Count == 0) return greedy;
-                    return MergeResults(greedy, GreedyAssignmentService.Assign(tailFiltered, greedy.Leftovers, pallet));
-                }, ct);
-
-                ct.ThrowIfCancellationRequested();
-
-                AssignmentResult? cpsatResult = null;
-                try
-                {
-                    cpsatResult = await Task.Run(() =>
-                    {
-                        ct.ThrowIfCancellationRequested();
-                        var filtered = LayerMetricsCalculator.FilterLayers(layersSnapshot, options);
-                        var pool = PalletTemplateEnumerator.Enumerate(pallet, filtered);
-                        var pruned = TemplateFilter.Filter(pool, pallet);
-                        return CPSATAssignmentService.Assign(pruned, demand, pallet, options, greedyResult, ct);
-                    }, ct);
-                }
-                catch (OperationCanceledException) { throw; }
-                catch { /* CP-SAT is best-effort; greedy result still shown */ }
-
-                ct.ThrowIfCancellationRequested();
-
-                Solutions.Clear();
-                if (cpsatResult != null && cpsatResult.Assignments.Count > 0)
-                    Solutions.Add(new SolutionDisplay(Solutions.Count + 1, "CP-SAT", cpsatResult, skus, _palletLength, _palletWidth, _palletHeight));
-                if (greedyResult.Assignments.Count > 0)
-                    Solutions.Add(new SolutionDisplay(Solutions.Count + 1, "Greedy", greedyResult, skus, _palletLength, _palletWidth, _palletHeight));
-
-                HasResults = Solutions.Count > 0;
-                if (HasResults)
-                {
-                    SelectedSolution = Solutions[0];
-                    PalletGenStats = $"Generated {Solutions.Count} solutions.";
-                    _snackbarService.Show("Generation complete",
-                        $"{_availableLayers.Count} candidate layers and {Solutions.Count} final {(Solutions.Count == 1 ? "solution" : "solutions")} generated.",
-                        ControlAppearance.Success, null, TimeSpan.FromSeconds(5));
-                }
-                else
-                    PalletGenStats = "No assignments produced.";
-            }
-            catch (OperationCanceledException) { }
-            catch (Exception ex)
-            {
-                PalletGenStats = $"Error: {ex.Message}";
-            }
-            finally
-            {
-                IsBuilding = false;
-            }
-        }
-
-        partial void OnSelectedSolutionChanged(SolutionDisplay? value)
-        {
-            Assignments.Clear();
-
-            if (value == null)
-            {
-                SelectedAssignment = null;
-                return;
-            }
-
-            int idx = 1;
-            foreach (var (template, count) in value.Result.Assignments)
-                Assignments.Add(new TemplateAssignmentDisplay(template, count, value.Skus, idx++, value.PalletLength, value.PalletWidth, value.PalletHeight));
-
-            SelectedAssignment = Assignments.FirstOrDefault();
-        }
-
-        partial void OnSelectedLayerTypeChanged(LayerTypeDisplay? value)
-        {
-            foreach (var h in _layerHighlights)
-                Scene.Children.Remove(h);
-            _layerHighlights.Clear();
-
-            if (value == null) return;
-
-            double inflate = 0.1;
-            var fillBrush = new SolidColorBrush(Color.FromArgb(40, 255, 255, 0));
-
-            foreach (var (layerId, y, height) in _sceneBuilder.LayerPositions)
-            {
-                if (layerId != value.LayerId) continue;
-                var origin = new Point3D(-inflate / 2.0, y - inflate / 2.0, -inflate / 2.0);
-                var highlight = GeometryCreator.CreateBoxWithEdges(
-                    origin,
-                    _palletLength + inflate,
-                    height + inflate,
-                    _palletWidth + inflate,
-                    fillBrush,
-                    Colors.Yellow,
-                    0.6);
-                Scene.Children.Add(highlight);
-                _layerHighlights.Add(highlight);
-            }
-        }
-
-        public bool TryGetLayerTypeForGeometry(GeometryModel3D geo, out LayerTypeDisplay? layerType)
-        {
-            layerType = null;
-            if (!_sceneBuilder.TryGetLayerIdForGeometry(geo, out var layerId)) return false;
-            layerType = SelectedLayerTypes.FirstOrDefault(lt => lt.LayerId == layerId);
-            return layerType != null;
-        }
-
-        partial void OnSelectedAssignmentChanged(TemplateAssignmentDisplay? value)
-        {
-            SelectedLayerType = null;
-
-            SelectedLayerTypes = value != null
-                ? new ObservableCollection<LayerTypeDisplay>(value.LayerTypes)
-                : [];
-
-            AssignmentDetailsText = value != null ? BuildAssignmentText(value) : string.Empty;
-
-            if (value != null && _viewportController != null)
-            {
-                double totalHeight = _palletHeight + value.Template.TotalHeight;
-                var center = new Point3D(_palletLength / 2.0, totalHeight / 2.0, _palletWidth / 2.0);
-                double distance = Math.Sqrt(_palletLength * _palletLength + _palletWidth * _palletWidth + totalHeight * totalHeight) * 1.5;
-                _viewportController.ResetView(center, distance);
-            }
-
-            _sceneCts?.Cancel();
-            _sceneCts?.Dispose();
-            _sceneCts = new CancellationTokenSource();
-            var ct = _sceneCts.Token;
-
-            if (value != null)
-                _ = UpdateSceneAsync(value.Template, ct);
-            else
-                Scene.Children.Clear();
-        }
-
-        private async Task UpdateSceneAsync(PalletTemplate template, CancellationToken ct)
-        {
-            try
-            {
-                await _sceneBuilder.BuildAsync(Scene, template, _palletLength, _palletWidth, _palletHeight, ct);
-                OnPropertyChanged(nameof(PalletDimLabels));
-            }
-            catch (OperationCanceledException) { }
-        }
-
-        private static AssignmentResult MergeResults(AssignmentResult main, AssignmentResult extra)
-        {
-            var merged = main.Assignments.ToList();
-            foreach (var (template, count) in extra.Assignments)
-            {
-                var idx = merged.FindIndex(a => a.Template.Id == template.Id);
-                if (idx >= 0)
-                    merged[idx] = (template, merged[idx].Count + count);
-                else
-                    merged.Add((template, count));
-            }
-            return new AssignmentResult { Assignments = merged, Leftovers = extra.Leftovers };
-        }
-
-        private static List<SKU> BuildLeftoverSkus(List<SKU> originalSkus, IReadOnlyDictionary<string, int> leftovers)
-        {
-            return [.. originalSkus
-                .Select(s => (Sku: s, Rem: leftovers.GetValueOrDefault(s.SkuId)))
-                .Where(x => x.Rem > 0)
-                .Select(x => new SKU
-                {
-                    SkuId = x.Sku.SkuId,
-                    Name = x.Sku.Name,
-                    Length = x.Sku.Length,
-                    Width = x.Sku.Width,
-                    Height = x.Sku.Height,
-                    Weight = x.Sku.Weight,
-                    Rotatable = x.Sku.Rotatable,
-                    Notes = x.Sku.Notes,
-                    Quantity = x.Rem
-                })];
-        }
-
-        private string BuildAssignmentText(TemplateAssignmentDisplay assignment)
-        {
-            var t = assignment.Template;
-            int cargoHeight = (int)Math.Round(t.TotalHeight);
-            int totalHeight = _palletHeight + cargoHeight;
-
-            var sb = new StringBuilder();
-            sb.AppendLine($"{assignment.Name}  (x{assignment.Count} {(assignment.Count == 1 ? "pallet" : "pallets")})");
-            sb.AppendLine();
-            sb.AppendLine($"Load dimensions: {_palletLength}x{_palletWidth}x{totalHeight} cm");
-            sb.AppendLine($"Cargo height: {cargoHeight} cm");
-            sb.AppendLine($"Weight: {t.TotalWeight:N0} kg");
-            sb.AppendLine($"Boxes: {t.TotalBoxCount} per pallet");
-            sb.AppendLine($"Utilization: {t.AverageLayerUtilization:F3}");
-            sb.AppendLine($"Contents: {assignment.Contents}");
-            sb.AppendLine("==================");
-            sb.AppendLine("Full details are included in the PDF report.");
-            return sb.ToString();
-        }
-    }
-
-    public class TemplateAssignmentDisplay
-    {
-        public PalletTemplate Template { get; }
-        public int Count { get; }
-        public string Name { get; }
-        public string SkuSummary { get; }
-        public string Contents { get; }
-        public string LoadDimensions { get; }
-        public string Weight { get; }
-        public string Efficiency { get; }
-        public IReadOnlyList<LayerTypeDisplay> LayerTypes { get; }
-
-        public TemplateAssignmentDisplay(PalletTemplate template, int count, IEnumerable<SKU> skuLookup, int index, int palletLength, int palletWidth, int palletHeight)
-        {
-            Template = template;
-            Count = count;
-            Name = $"Type {index}";
-            Efficiency = template.AverageLayerUtilization.ToString("P0");
-            Weight = template.TotalWeight.ToString("N0");
-            LoadDimensions = $"{palletLength}x{palletWidth}x{(int)Math.Round(palletHeight + template.TotalHeight)}";
-
-            var skuMap = skuLookup.ToDictionary(s => s.SkuId, s => s.Name, StringComparer.Ordinal);
-            SkuSummary = string.Join("  |  ", template.SkuCounts
-                .OrderBy(kvp => skuMap.GetValueOrDefault(kvp.Key, kvp.Key), StringComparer.Ordinal)
-                .Select(kvp => $"{skuMap.GetValueOrDefault(kvp.Key, kvp.Key)} ×{kvp.Value}"));
-            Contents = string.Join(", ", template.SkuCounts
-                .OrderBy(kvp => skuMap.GetValueOrDefault(kvp.Key, kvp.Key), StringComparer.Ordinal)
-                .Select(kvp => $"{kvp.Value}x {skuMap.GetValueOrDefault(kvp.Key, kvp.Key)}"));
-            LayerTypes = [.. template.Layers
-                .GroupBy(l => l.Id)
-                .Select(g => new LayerTypeDisplay(g.First(), g.Count(), skuMap))];
-        }
-    }
-
-    public class LayerTypeDisplay(Layer layer, int count, IReadOnlyDictionary<string, string> skuNames)
-    {
-        public string LayerId { get; } = layer.Id;
-        public string Name { get; } = layer.Name;
-        public int Count { get; } = count;
-        public string Contents { get; } = string.Join(", ", layer.Items
-                .GroupBy(i => i.SkuType.SkuId)
-                .Select(g => $"{g.Count()}x {skuNames.GetValueOrDefault(g.Key, g.Key)}"));
-        public string Utilization { get; } = layer.Metadata.Utilization.ToString("P0");
-    }
-
-    public partial class SolutionDisplay : ObservableObject
-    {
-        public int Number { get; }
-        public string Name { get; }
-        public AssignmentResult Result { get; }
-        public IReadOnlyList<SKU> Skus { get; }
-        public int PalletLength { get; }
-        public int PalletWidth { get; }
-        public int PalletHeight { get; }
-        public int TotalPallets { get; }
-        public int PalletTypes { get; }
-        public int TotalItemsPacked { get; }
-        public string Efficiency { get; }
-
-        public SolutionDisplay(
-            int number,
-            string name,
-            AssignmentResult result,
-            IReadOnlyList<SKU> skus,
-            int palletLength,
-            int palletWidth,
-            int palletHeight)
-        {
-            Number = number;
-            Name = name;
-            Result = result;
-            Skus = skus;
-            PalletLength = palletLength;
-            PalletWidth = palletWidth;
-            PalletHeight = palletHeight;
-            TotalPallets = result.TotalPallets;
-            PalletTypes = result.Assignments.Count;
-            TotalItemsPacked = result.Assignments.Sum(a => a.Template.TotalBoxCount * a.Count);
-            var weightedUtil = result.TotalPallets > 0
-                ? result.Assignments.Sum(a => a.Template.AverageLayerUtilization * a.Count) / result.TotalPallets
-                : 0;
-            Efficiency = weightedUtil.ToString("P0");
-        }
-    }
-}
