+++ ViewModels/Pages/PalletBuilderViewModel.cs
@@ -21,8 +21,7 @@ namespace Stack_Solver.ViewModels.Pages
         ISnackbarService snackbarService) : ObservableObject
     {
         public PalletBuilderSettingsViewModel Settings { get; } = new PalletBuilderSettingsViewModel(skuRepository, events, genOptions, palletDefaults, settingsValidator, skuQuantityValidator, userSettings);
-        public LayerAnalyzerViewModel LayerAnalyzer { get; } = new LayerAnalyzerViewModel(events, viz, snackbarService);
-        public PalletAnalyzerViewModel PalletAnalyzer { get; } = new PalletAnalyzerViewModel(events, snackbarService);
+        public ResultsViewModel Results { get; } = new ResultsViewModel(events, viz, snackbarService);
 
         public async Task OnNavigatedToAsync()
         {
